fix prototype for __checked_l()

+++ __ung.h
@@ -182,7 +182,7 @@
 
 void __checked(const char *, const char *, unsigned long long, void(*)(), ...);
 int __checked_i(const char *, const char *, unsigned long long, int(*)(), ...);
-long __checked_l(const char *, const char *, unsigned long long, int(*)(), ...);
+long __checked_l(const char *, const char *, unsigned long long, long(*)(), ...);
 void * __checked_p(const char *, const char *, unsigned long long, void *(*)(), ...);
 
 #if (!defined __STDC_VERSION__) || (__STDC_VERSION__ < 199901L)
